SAME Asymptotic Variance

Calculates the asymptotic variance (or variance - covariance matrix in the multidimensional case) of the SAME, given a specified family of distributions and the true parameter values.

Usage

avar_same(distr, ...)

# S4 method for Beta
avar_same(distr)

# S4 method for Dirichlet
avar_same(distr, comp = FALSE)

# S4 method for Gammad
avar_same(distr)

# S4 method for MGamma
avar_same(distr, dirich = FALSE)

Arguments

distr

A subclass of Distribution. The distribution family assumed.

...

extra arguments.

comp

logical. Should the component matrices A, B be returned instead? This argument is used internally by the avar_same method for the Multivariate Gamma Dirichlet-based SAME that needs these matrices.

dirich

logical. Should the Dirichlet-based estimator be calculated instead? Applies only to the Multivariate Gamma distribution.

Value

A named matrix. The asymptotic covariance matrix of the estimator.

See also

Examples

if (FALSE) {
# -------------------------------------------
# Beta Distribution Example
# -------------------------------------------

# Simulation

set.seed(1)
x <- rbeta(100, shape1, shape2)
D <- Beta(shape1, shape2)

# Likelihood - The ll Functions

llbeta(x, shape1, shape2)
ll(x, c(shape1, shape2), D)
ll(x, c(shape1, shape2), "beta")

# Point Estimation - The e Functions

ebeta(x, type = "mle")
ebeta(x, type = "me")
ebeta(x, type = "same")

mle(x, D)
me(x, D)
same(x, D)

estim(x, D, type = "mle")

# Asymptotic Variance - The v Functions

vbeta(shape1, shape2, type = "mle")
vbeta(shape1, shape2, type = "me")
vbeta(shape1, shape2, type = "same")

avar_mle(D)
avar_me(D)
avar_same(D)

avar(D, type = "mle")
}