Moment Estimator

Calculates the MΕ of a sample under the assumption the observations are independent and identically distributed (iid) according to a specified family of distributions.

me(x, distr, ...)

# S4 method for ANY,character
me(x, distr, ...)

# S4 method for numeric,Beta
me(x, distr)

# S4 method for matrix,Beta
me(x, distr)

# S4 method for numeric,Gammad
me(x, distr)

# S4 method for matrix,Gammad
me(x, distr)

# S4 method for matrix,Dirichlet
me(x, distr)

# S4 method for matrix,MGamma
me(x, distr)

Arguments

x

numeric. A sample under estimation.

distr

A subclass of Distribution. The distribution family assumed.

...

extra arguments.

Value

numeric or list. The estimator produced by the sample. In case the distribution parameters hold specific structures (e.g. the Matrix Gamma parameters are a positive real number and a positive definite matrix), a list is returned instead.

Examples

if (FALSE) {
# Distribution
d_beta <- Beta(shape1 = 1, shape2 = 1.5)

# Simulation
set.seed(2)
x <- r(d_beta)(50)

# Estimation
me(x, "Beta")
same(x, "Beta")
mle(x, "Beta")

# Asymptotic Covariance Matrix
acov_me(d_beta)
acov_same(d_beta)
acov_mle(d_beta)
}