Calculate Metrics

This function performs Monte Carlo simulations to estimate the main metrics (bias, variance, and RMSE) characterizing the finite sample behavior of an estimator. The function evaluates the metrics as a function of a single parameter, keeping the other ones constant. See Details.

Usage

metrics(
  D0,
  prm,
  obs = c(20, 50, 100),
  est = c("same", "me", "mle"),
  sam = 10000,
  seed = 1
)

Arguments

D0

A subclass of Distribution. The distribution family of interest.

prm

A list containing three elements (name, pos, val). See Details.

obs

numeric. The size of each sample. Can be a vector.

est

character. The estimator of interest. Can be a vector.

sam

numeric. The number of Monte Carlo samples used to estimate the metrics.

seed

numeric. Passed to set.seed() for reproducibility.

Value

A data.frame with columns named "Parameter", "Observations", "Estimator", "Metric", and "Value".

Details

The distribution D0 is used to specify an initial distribution. The list prm contains details concerning a single parameter that is allowed to change values. The quantity of interest is evaluated as a function of this parameter.

Specifically, prm includes three elements named "name", "pos", and "val". The first two elements determine the exact parameter that changes, while the third one is a numeric vector holding the values it takes. For example, in the case of the Multivariate Gamma distribution, D0 <- MGamma(shape = c(1, 2), scale = 3) and prm <- list(name = "shape", pos = 2, val = seq(1, 1.5, by = 0.1)) means that the evaluation will be performed for the MGamma distributions with shape parameters (1, 1), (1, 1.1), ..., (1, 1.5) and scale 3. Notice that the initial shape parameter 2 in D0 is not utilized in the function.

See also

Examples

if (FALSE) {
# Distribution
D <- Beta(shape1 = 1, shape2 = 1.5)

# Simulation
set.seed(2)
x <- r(D)(50)

# Estimation
me(x, "Beta")
same(x, "Beta")
mle(x, "Beta")

# Asymptotic Covariance Matrix
acov_me(D)
acov_same(D)
acov_mle(D)
}