Moments - Parametric Quantities of Interest

A set of functions that calculate the theoretical moments (expectation, variance, skewness, excess kurtosis) and other important parametric functions (median, mode, entropy, Fisher information) of a distribution.

Usage

moments(x)

mean(x, ...)

median(x, na.rm = FALSE, ...)

mode(x)

var(x, y = NULL, na.rm = FALSE, use)

sd(x, na.rm = FALSE)

skew(x, ...)

kurt(x, ...)

entro(x, ...)

finf(x, ...)

Arguments

x

an object of a Distribution subclass.

...

extra arguments.

y, use, na.rm

arguments in mean and var standard methods from the stats package not used here.

Value

Numeric, either vector or matrix depending on the moment and the distribution. Function moments() returns a list of all available methods.

Details

The moments() function automatically finds the available methods for a given distribution and results all of the results in a list.

Not all functions are available for distributions; for example, the sd() is available only for univariate distributions.

Examples

# -----------------------------------------------------
# Beta Distribution Example
# -----------------------------------------------------

library(estimators)

# Create the distribution
x <- Beta(3, 5)

# List of all available moments
mom <- moments(x)

# Expectation
mean(x)
#> [1] 0.375
mom$mean
#> [1] 0.375

# Variance and Standard Deviation
var(x)
#> [1] 0.02604167
sd(x)
#> [1] 0.1613743

# Skewness and Excess Kurtosis
skew(x)
#> [1] 0.3098387
kurt(x)
#> [1] 0.04

# Entropy
entro(x)
#> [1] -0.4301508

# Fisher Information Matrix
finf(x)
#>            shape1      shape2
#> shape1  0.2617971 -0.13313701
#> shape2 -0.1331370  0.08818594